backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
scikit-learn
pydantic
python-dotenv
alembic
